WEB-3237: more fixes after merge
