+++ src/authentication/decorators/PublicAccess.decorator.ts
@@ -1,4 +1,5 @@
 import { SetMetadata } from '@nestjs/common'
 
 export const IS_PUBLIC_KEY = 'isPublic'
+/** Tells JwtAuthGuard to skip auth for the controller/handler this decorator is applied to */
 export const PublicAccess = () => SetMetadata(IS_PUBLIC_KEY, true)
