+++ src/authentication/schemas/ResetPassword.schema.ts
@@ -1,13 +1,13 @@
 import { createZodDto } from 'nestjs-zod'
+import { PasswordSchema } from 'src/users/schemas/Password.schema'
 import { z } from 'zod'
-import { passwordSchema } from '../../users/util/passwords.util'
 
 export class ResetPasswordSchema extends createZodDto(
   z
     .object({
       email: z.string().email(),
       token: z.string(),
-      password: passwordSchema,
+      password: PasswordSchema,
       adminCreate: z.boolean().optional(),
     })
     .strict(),
