+++ src/users/users.module.ts
@@ -1,7 +1,9 @@
 import { Module } from '@nestjs/common'
 import { UsersService } from './users.service'
+import { UsersController } from './users.controller'
 
 @Module({
+  controllers: [UsersController],
   providers: [UsersService],
   exports: [UsersService],
 })
